+++ springboot-web-mvc/src/main/java/com/learning/springboot/bootwebmvc/part01_web_mvc/UserController.java
@@ -20,4 +20,10 @@ public class UserController {
         long id = idSequence.incrementAndGet();
         return new UserResponse(id, request.getName(), request.getEmail());
     }
+
+    @PostMapping("/no-valid")
+    public UserResponse createUserWithoutValid(@RequestBody CreateUserRequest request) {
+        long id = idSequence.incrementAndGet();
+        return new UserResponse(id, request.getName(), request.getEmail());
+    }
 }